agent/schedwatch: Split up sched.go
